seperate code into Export.hs & Gen.hs
